% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allin}
\alias{allin}
\title{collection of synthea FHIR documents ingested}
\format{
list
}
\usage{
allin
}
\description{
collection of synthea FHIR documents ingested
}
\examples{
data("allin", package="BiocFHIR")
allin[[1]]
}
\keyword{datasets}
