% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_json.R
\name{available_retention_schemas}
\alias{available_retention_schemas}
\title{list available 'retention schemas'}
\usage{
available_retention_schemas()
}
\value{
character vector
}
\description{
list available 'retention schemas'
}
\examples{
available_retention_schemas()
}
