% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphit.R
\name{display_proccond_igraph}
\alias{display_proccond_igraph}
\title{make network visualization}
\usage{
display_proccond_igraph(igraph)
}
\description{
make network visualization
}
\examples{
data(allin)
g = build_proccond_igraph( allin ) 
if (interactive()) {
 display_proccond_igraph( g )
}
}
