% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{FHIR_ResourceTypes}
\alias{FHIR_ResourceTypes}
\title{FHIR Resource types recognized in package}
\usage{
FHIR_ResourceTypes()
}
\description{
FHIR Resource types recognized in package
}
\examples{
FHIR_ResourceTypes() # Oct 2022
}
