% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{FHIR_retention_schemas}
\alias{FHIR_retention_schemas}
\title{collection of FHIR Resource components to be retained}
\usage{
FHIR_retention_schemas()
}
\description{
collection of FHIR Resource components to be retained
}
\examples{
FHIR_retention_schemas() # Oct 2022
}
