% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_app.R
\name{freq_app}
\alias{freq_app}
\title{produce interactive tables with FHIR resources from a list of ingested bundles}
\usage{
freq_app(blist)
}
\arguments{
\item{blist}{list of ingested bundles}
}
\value{
side-effects of shiny app invocation
}
\description{
produce interactive tables with FHIR resources from a list of ingested bundles
}
