% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpat.R
\name{get_patient_demog}
\alias{get_patient_demog}
\title{process resources from a FHIR R4 json bundle to acquire selected patient demographics}
\usage{
get_patient_demog(reslist, reduce = TRUE)
}
\arguments{
\item{reslist}{output of `get_resources`}

\item{reduce}{logical(1) if TRUE a vector is returned, otherwise a Patient.parse_obj
value is passed back}
}
\description{
process resources from a FHIR R4 json bundle to acquire selected patient demographics
}
\examples{
example(get_resources)
}
