% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpat.R
\name{get_resources}
\alias{get_resources}
\title{process FHIR R4 json to acquire "Resources" from a single patient bundle}
\usage{
get_resources(jsonpath, maxres = 50)
}
\arguments{
\item{jsonpath}{character(1) path to json}

\item{maxres}{numeric(1) current resource processing is slow using self-appending, set maxres to limit for testing}
}
\value{
a list with two entries: `resources`, a 'py_eval' result with the python list of resources,
and `resources_dict` suitable for Patient.parse_obj from fhir.resources
}
\description{
process FHIR R4 json to acquire "Resources" from a single patient bundle
}
\examples{
demfile = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json", package="BiocFHIR")
chk = get_resources(demfile)
get_patient_demog(chk)
}
