% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphit.R
\name{make_condition_graph}
\alias{make_condition_graph}
\title{create graph with links from patients to conditions}
\usage{
make_condition_graph(listOfProcessedBundles)
}
\arguments{
\item{listOfProcessedBundles}{list with elements generated by `process_fhir_bundle`}
}
\description{
create graph with links from patients to conditions
}
