% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_json.R
\name{print.FHIR.bundle}
\alias{print.FHIR.bundle}
\title{print method}
\usage{
\method{print}{FHIR.bundle}(x, ...)
}
\arguments{
\item{x}{BiocFHIR FHIR.bundle instance}

\item{\dots}{not used}
}
\description{
print method
}
