% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOthers.R
\name{process_AllergyIntolerance}
\alias{process_AllergyIntolerance}
\title{extract information from retained fields in Observation component of FHIR Bundle, produce simple data.frame}
\usage{
process_AllergyIntolerance(AllergyIntolerance)
}
\arguments{
\item{AllergyIntolerance}{component of FHIR.bundle instance}
}
\value{
data.frame
}
\description{
extract information from retained fields in Observation component of FHIR Bundle, produce simple data.frame
}
\examples{
testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_AllergyIntolerance(tbun$AllergyIntolerance)
}
