% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_CarePlan.R
\name{process_CarePlan}
\alias{process_CarePlan}
\title{extract information from retained fields in CarePlan component of FHIR Bundle, produce simple data.frame}
\usage{
process_CarePlan(CarePlan)
}
\arguments{
\item{CarePlan}{component of FHIR.bundle instance}
}
\value{
data.frame
}
\description{
extract information from retained fields in CarePlan component of FHIR Bundle, produce simple data.frame
}
\examples{
testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_CarePlan(tbun$CarePlan)
}
