% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_Condition.R
\name{process_Condition}
\alias{process_Condition}
\title{extract information from retained fields in Condition component of FHIR Bundle, produce simple data.frame}
\usage{
process_Condition(Condition)
}
\arguments{
\item{Condition}{component of FHIR.bundle instance}
}
\value{
data.frame
}
\description{
extract information from retained fields in Condition component of FHIR Bundle, produce simple data.frame
}
\examples{
testf <- system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun <- process_fhir_bundle(testf)
process_Condition(tbun$Condition)
}
