% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_json.R
\name{selectSome}
\alias{selectSome}
\title{from Biobase ...}
\usage{
selectSome(obj, maxToShow = 5)
}
\arguments{
\item{obj}{any vector}

\item{maxToShow}{numeric(1)}
}
\value{
character vector
}
\description{
from Biobase ...
}
