% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_fhir.R
\name{stack_fhir}
\alias{stack_fhir}
\title{convert data of a given FHIR type in a list of bundles to a data.frame}
\usage{
stack_fhir(blist, type, droperr = TRUE)
}
\arguments{
\item{blist}{list of FHIR bundles imported with process_fhir_bundle}

\item{type}{character(1) type, in names(blist[[1]]), e.g.}

\item{droperr}{logical(1) exclude records for which process_[type] fails, defaults to TRUE}
}
\value{
data.frame
}
\description{
convert data of a given FHIR type in a list of bundles to a data.frame
}
\examples{
jj = make_test_json_set()
b2 = lapply(jj[1:2], process_fhir_bundle)
ss = stack_fhir(b2, "Procedure")
head(ss,2)
}
