% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{summarise_bundles}
\alias{summarise_bundles}
\title{produce tables summarizing FHIR data}
\usage{
summarise_bundles(
  blist,
  resource = "Condition",
  selection_map = summary_selections()
)
}
\arguments{
\item{blist}{list of ingested bundles}

\item{resource}{character(1) FHIR resource name}

\item{selection_map}{character() named vector of single strings selected for summarisation}
}
\value{
data.frame
}
\description{
produce tables summarizing FHIR data
}
