% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{summary_selections}
\alias{summary_selections}
\title{vector of fields to be selected for summarization}
\usage{
summary_selections()
}
\value{
named vector of strings
}
\description{
vector of fields to be selected for summarization
}
