% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphit.R
\name{add_procedures}
\alias{add_procedures}
\title{update a fhir graph on patients and conditions with procedures}
\usage{
add_procedures(fhirgraph, listOfProcessedBundles)
}
\description{
update a fhir graph on patients and conditions with procedures
}
