% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphit.R
\name{add_procedures}
\alias{add_procedures}
\title{update a fhir graph on patients and conditions with procedures}
\usage{
add_procedures(fhirgraph, listOfProcessedBundles)
}
\arguments{
\item{fhirgraph}{instance of BiocFHIR.FHIRgraph}

\item{listOfProcessedBundles}{list}
}
\value{
instance of BiocFHIR.FHIRgraph
}
\description{
update a fhir graph on patients and conditions with procedures
}
\examples{
data("allin")
g = make_condition_graph(allin)
g = add_procedures(g, allin)
g
}
