% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphit.R
\name{build_proccond_igraph}
\alias{build_proccond_igraph}
\title{build graph with patients, conditions and procedures}
\usage{
build_proccond_igraph(listOfBundles)
}
\arguments{
\item{listOfBundles}{list of processed FHIR bundles, processed with `process_fhir_bundle`}
}
\description{
build graph with patients, conditions and procedures
}
