% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphit.R
\name{make_condition_graph}
\alias{make_condition_graph}
\title{create graph with links from patients to conditions}
\usage{
make_condition_graph(listOfProcessedBundles, keep_with_condition_only = TRUE)
}
\arguments{
\item{listOfProcessedBundles}{list with elements generated by `process_fhir_bundle`}

\item{keep_with_condition_only}{logical(1) omit bundles that lack a "Condition" element, defaults to TRUE}
}
\value{
instance of BiocFHIR.FHIRgraph
}
\description{
create graph with links from patients to conditions
}
\examples{
data("allin")
make_condition_graph(allin, TRUE)
}
