% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_json.R
\name{process_fhir_bundle}
\alias{process_fhir_bundle}
\title{process a bundle of FHIR R4 JSON}
\usage{
process_fhir_bundle(json_file)
}
\arguments{
\item{json_file}{character(1) path to text in JSON format}
}
\value{
instance of FHIR.bundle, extending list
}
\description{
process a bundle of FHIR R4 JSON
}
\examples{
testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
tbun
}
